requirements.txt
# Core dependencies
asyncio>=3.4.3
typing-extensions>=4.0.0  # Latest stable for Python 3.9+
dataclasses>=0.6  # Latest stable
click>=8.0.0  # Latest stable with LTS support
PyYAML>=6.0.1

# Async and networking
aiohttp>=3.8.0  # Latest stable with security updates
anyio>=3.0.0  # Latest stable

# Data processing and validation
numpy>=1.21.0  # Latest stable with wide compatibility
pydantic>=1.8.2  # Major version update with improvements

# Development and testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
mypy>=0.910
flake8>=3.9.2
black>=21.7b0

# Documentation
sphinx>=7.2.6  # Optional, for building docs
sphinx-rtd-theme>=1.3.0  # Optional, for documentation theme
